ref(monitors): Inline error breadcrumb variants
Keep the error monitor detail and edit breadcrumbs local to their pages instead of sharing a helper.

The detail and edit paths now own their slightly different breadcrumb behavior directly.
