+++ static/app/views/detectors/components/forms/common/breadcrumbs.tsx
@@ -1,7 +1,5 @@
 import {Breadcrumbs} from 'sentry/components/breadcrumbs';
-import ProjectBadge from 'sentry/components/idBadge/projectBadge';
 import {t} from 'sentry/locale';
-import type {Project} from 'sentry/types/project';
 import type {Detector, DetectorType} from 'sentry/types/workflowEngine/detectors';
 import {useOrganization} from 'sentry/utils/useOrganization';
 import {
@@ -50,44 +48,6 @@ export function EditDetectorBreadcrumbs({detector}: {detector: Detector}) {
           to: makeMonitorDetailsPathname(organization.slug, detector.id),
         },
         {label: t('Configure')},
-      ]}
-    />
-  );
-}
-
-export function ErrorDetectorProjectBreadcrumbs({
-  detector,
-  project,
-  includeConfigure = false,
-}: {
-  detector: Detector;
-  includeConfigure?: boolean;
-  project?: Project;
-}) {
-  const organization = useOrganization();
-
-  return (
-    <Breadcrumbs
-      crumbs={[
-        {
-          label: t('Monitors'),
-          to: makeMonitorBasePathname(organization.slug),
-        },
-        {
-          label: getDetectorTypeLabel(detector.type),
-          to: makeMonitorTypePathname(organization.slug, detector.type),
-        },
-        ...(project
-          ? [
-              {
-                label: <ProjectBadge disableLink project={project} avatarSize={16} />,
-                ...(includeConfigure
-                  ? {to: makeMonitorDetailsPathname(organization.slug, detector.id)}
-                  : {}),
-              },
-            ]
-          : []),
-        ...(includeConfigure ? [{label: t('Configure')}] : []),
       ]}
     />
   );
