+++ static/app/views/detectors/components/forms/error/index.tsx
@@ -8,23 +8,29 @@ import {Stack} from '@sentry/scraps/layout';
 import {ExternalLink} from '@sentry/scraps/link';
 import {Text} from '@sentry/scraps/text';
 
+import {Breadcrumbs} from 'sentry/components/breadcrumbs';
 import {FormContext} from 'sentry/components/forms/formContext';
+import ProjectBadge from 'sentry/components/idBadge/projectBadge';
 import * as Layout from 'sentry/components/layouts/thirds';
 import {LoadingError} from 'sentry/components/loadingError';
 import {EditLayout} from 'sentry/components/workflowEngine/layout/edit';
 import {Container} from 'sentry/components/workflowEngine/ui/container';
 import {FormSection} from 'sentry/components/workflowEngine/ui/formSection';
 import {t, tct} from 'sentry/locale';
+import type {Project} from 'sentry/types/project';
 import type {ErrorDetector} from 'sentry/types/workflowEngine/detectors';
 import {useOrganization} from 'sentry/utils/useOrganization';
 import {useProjectFromId} from 'sentry/utils/useProjectFromId';
 import {AutomationFeedbackButton} from 'sentry/views/automations/components/automationFeedbackButton';
 import {AutomateSection} from 'sentry/views/detectors/components/forms/automateSection';
-import {
-  EditDetectorBreadcrumbs,
-  ErrorDetectorProjectBreadcrumbs,
-} from 'sentry/views/detectors/components/forms/common/breadcrumbs';
+import {EditDetectorBreadcrumbs} from 'sentry/views/detectors/components/forms/common/breadcrumbs';
 import {useEditDetectorFormSubmit} from 'sentry/views/detectors/hooks/useEditDetectorFormSubmit';
+import {
+  makeMonitorBasePathname,
+  makeMonitorDetailsPathname,
+  makeMonitorTypePathname,
+} from 'sentry/views/detectors/pathnames';
+import {getDetectorTypeLabel} from 'sentry/views/detectors/utils/detectorTypeConfig';
 import {getNoPermissionToEditMonitorTooltip} from 'sentry/views/detectors/utils/monitorAccessMessages';
 import {useCanEditDetectorWorkflowConnections} from 'sentry/views/detectors/utils/useCanEditDetector';
 import {TopBar} from 'sentry/views/navigation/topBar';
@@ -120,6 +126,40 @@ export function NewErrorDetectorForm() {
   );
 }
 
+function ErrorEditBreadcrumbs({
+  detector,
+  project,
+}: {
+  detector: ErrorDetector;
+  project?: Project;
+}) {
+  const organization = useOrganization();
+
+  return (
+    <Breadcrumbs
+      crumbs={[
+        {
+          label: t('Monitors'),
+          to: makeMonitorBasePathname(organization.slug),
+        },
+        {
+          label: getDetectorTypeLabel(detector.type),
+          to: makeMonitorTypePathname(organization.slug, detector.type),
+        },
+        ...(project
+          ? [
+              {
+                label: <ProjectBadge disableLink project={project} avatarSize={16} />,
+                to: makeMonitorDetailsPathname(organization.slug, detector.id),
+              },
+            ]
+          : []),
+        {label: t('Configure')},
+      ]}
+    />
+  );
+}
+
 export function EditExistingErrorDetectorForm({detector}: {detector: ErrorDetector}) {
   const project = useProjectFromId({project_id: detector.projectId});
   const theme = useTheme();
@@ -157,11 +197,7 @@ export function EditExistingErrorDetectorForm({detector}: {detector: ErrorDetect
       {hasPageFrameFeature ? (
         <Fragment>
           <TopBar.Slot name="title">
-            <ErrorDetectorProjectBreadcrumbs
-              detector={detector}
-              project={project}
-              includeConfigure
-            />
+            <ErrorEditBreadcrumbs detector={detector} project={project} />
           </TopBar.Slot>
           <AutomationFeedbackButton />
         </Fragment>
